Allow elements to be ridiculously customizable
Split the input element into separate input and label elements, and export a
new type called labeledInput to be more truthful to what you're getting.

Allow classes and styles to be added (open q: should we allow removing current?,
and if so, how?), and allow arbitrary attributes to be set on the elements
allowing for custom behaviors.